Complete async GZIP implementation
It turns out this implementation is EXTREMELY slow because of its
excessive use of await. If we really want to still do this, we need
to process data in blocking chunks rather than continuous + async.

An idea as of now:
1. Throw an error from the stream reader when no data is available.
2. Catch the error in the block reading loop.
3. Allows the reading buffer to catch up.
4. Parse block from the start (rewind stream reader).

Also, we might want to use typed arrays for buffering in the stream
reader. We could hold a list of chunks as typed arrays and trash
an entire chunk at a time.